+++ lib/flu-rails/errors.rb
@@ -1,0 +1,9 @@
+# frozen_string_literal: true
+
+module Flu
+  class Error < StandardError
+  end
+
+  class NotConnectedError < Error
+  end
+end